Add Usage/Help output for main app and subcommands
- Display help if explicitly requested
- Display help if subcommands not provided

This implementation is a mix of static subcommand
names and dynamic help generation. While this set
of changes does work, it does not feel as clean or
maintainable as I would like.

I hope to return to this later once I learn more
about Go and implement this in a more consistent
manner.

refs #31